refactored CellSettings to use the common view
